requirements.txt
torch
torchvision
opencv-python
numpy
pandas
matplotlib
seaborn
scikit-learn
streamlit
PyYAML
gdown
tqdm
